% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_flopy.R
\name{import_flopy}
\alias{import_flopy}
\title{Import "flopy" Python Package}
\usage{
import_flopy(convert = TRUE, ...)
}
\arguments{
\item{convert}{Boolean (default: TRUE); should Python objects be automatically
converted to their R equivalent? If set to FALSE, you can still manually convert
Python objects to R via the \link[reticulate]{py_to_r} function.}

\item{...}{additional arguments passed to \link[reticulate]{import}}
}
\value{
imports "fl
}
\description{
Import "flopy" Python Package
}
