% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{plot_data}
\alias{plot_data}
\title{plot_data}
\usage{
plot_data(
  multiarray,
  title = "",
  value_min = NULL,
  value_max = NULL,
  fill_gradient_low = "grey90",
  fill_gradient_high = "red"
)
}
\arguments{
\item{multiarray}{multidimensional MODFLOW array}

\item{title}{title for plot}

\item{value_min}{minimum value. All smaller values will be excluded from dataset
(default: NULL)}

\item{value_max}{maximum value. All larger values will be excluded from dataset
(default: NULL)}

\item{fill_gradient_low}{fill_gradient_low (default: "grey90")}

\item{fill_gradient_high}{fill_gradient_high (default: "red")}
}
\value{
plot all Modflow layers
}
\description{
plot_data
}
