% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{to_long}
\alias{to_long}
\title{Helper function: convert multiarray to long format}
\usage{
to_long(multiarray)
}
\arguments{
\item{multiarray}{multidimensional MODFLOW array}
}
\value{
convert multiarray to long format with columns:
layer, column, row, value
}
\description{
Helper function: convert multiarray to long format
}
