% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_wide.R
\name{to_wide}
\alias{to_wide}
\title{to_wide}
\usage{
to_wide(multiarray_long, parameter = "")
}
\arguments{
\item{multiarray_long}{multiarray in long format (as retrieved by
\code{\link{to_long}})}

\item{parameter}{prefix to use for parameter in wide format (default: "")}
}
\value{
tibble with column, row, parameter_layer1-n
}
\description{
to_wide
}
