% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{plot_data}
\alias{plot_data}
\title{plot_data}
\usage{
plot_data(mulitarray, title = "")
}
\arguments{
\item{mulitarray}{multidimensional MODFLOW array}

\item{title}{title for plot}
}
\value{
plot all Modflow layers
}
\description{
plot_data
}
