% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{plot_data}
\alias{plot_data}
\title{plot_data}
\usage{
plot_data(mulitarray, title = "", value_min = NULL, value_max = NULL)
}
\arguments{
\item{mulitarray}{multidimensional MODFLOW array}

\item{title}{title for plot}

\item{value_min}{minimum value. All smaller values will be excluded from dataset
(default: NULL)}

\item{value_max}{maximum value. All larger values will be excluded from dataset
(default: NULL)}
}
\value{
plot all Modflow layers
}
\description{
plot_data
}
